src/requirements.txt
Flask
numpy
pandas
pillow
opencv-python
tensorflow_hub
